requirements.txt
moviepy>=1.0.3
pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
pydub>=0.25.1
# Core video processing
moviepy
opencv-python
numpy
pillow

# AI/ML for facial animation
torch
torchvision


# Audio processing
librosa
scipy
mysqlclient

# Additional utilities
requests
gdown  # for downloading models